feat(janus_pro): pathch_embedding file renaming
